### 2019-06-06 #1490 @whedon
[PRE REVIEW]: phasespace: $n$-body phase space generation in Python
**Submitting author:** @apuignav (<a href="http://orcid.org/0000-0001-8868-2947">Albert Puig Navarro</a>)
**Repository:** <a href="https://github.com/zfit/phasespace/" target ="_blank">https://github.com/zfit/phasespace/</a>
**Version:** 1.0.2
