### 2019-06-06 #1490 @labarba
@apuignav Please check your `.bib` file... I think you're missing a comma in line 12.
