+++ paper/paper.bib
@@ -9,7 +9,7 @@
       year           = "1997",
       pages          = "81-86",
       doi            = "10.1016/S0168-9002(97)00048-X",
-      notes          = "{See also \url{http://root.cern.ch/}}"
+      notes          = "{See also \url{http://root.cern.ch/}}",
       SLACcitation   = "%%CITATION = NUIMA,A389,81;%%"
 }
 
