### 2019-06-06 #1490 @apuignav
> @apuignav Please check your `.bib` file... I think you're missing a comma in line 12.

Yup, sorry, done!
