### 2019-06-06 2a9813c
Fix small mistakes after proofs.
