+++ paper/paper.bib
@@ -98,7 +98,7 @@ author={
 
 
 @misc{zfit,
-	Author = {Jonas Eschle and Albert Puig Navarro and Rafael Silva Coutinho},
+	Author = {Jonas Eschle and Albert {Puig Navarro} and Rafael {Silva Coutinho}},
 	Date-Added = {2019-03-12 14:59:41 +0100},
 	Date-Modified = {2019-03-21 14:39:21 -0400},
 	Doi = {10.5281/zenodo.2602043},
@@ -110,6 +110,6 @@ author={
  title={phasespace: n-body phase space generation in Python},
  DOI={10.5281/zenodo.2591993},
  publisher={Zenodo},
- author={Albert Puig and Jonas Eschle},
+ author={Albert {Puig Navarro} and Jonas Eschle},
  year={2019},
  month={Mar}}
