### 2019-06-06 #1490 @apuignav
Latest proof looks good to me!
