### 2019-06-09 #1490 @poulson
:wave: Tim (@betatim), it seems that your expertise in physics and python would be helpful for reviewing this submission. Would you be willing and able to do so?

:wave: Pawel (@jochym), given your expertise in theoretical physics and python, your help reviewing this submission would be greatly appreciated. Are you available and interested?

:wave: Mathieu (@mdoucet), given your expertise in neutron scattering and python, your thoughts on this submission would be immensely helpful. Are you interested and available?
