### 2019-06-12 9bf2caf
Add mass info in repr
