+++ phasespace/phasespace.py
@@ -91,6 +91,7 @@ class Particle:
     def __init__(self, name: str, mass: Union[Callable, int, float]) -> None:  # noqa
         self.name = name
         self.children = []
+        self._mass_val = mass
         if not callable(mass) and not isinstance(mass, tf.Variable):
             mass = tf.convert_to_tensor(mass, preferred_dtype=tf.float64)
             mass = tf.cast(mass, tf.float64)
@@ -99,8 +100,11 @@ class Particle:
         self._cache = None
 
     def __repr__(self):
-        return "<phasespace.Particle: name='{}' children=[{}]>".format(self.name,
-                                                                       ', '.join(child.name for child in self.children))
+        return "<phasespace.Particle: name='{}' mass={} children=[{}]>".format(self.name,
+                                                                               f"{self._mass_val:.2f}"
+                                                                               if self.has_fixed_mass
+                                                                               else "variable",
+                                                                               ', '.join(child.name for child in self.children))
 
     @property
     def _sess(self):
