### 2019-06-14 2747618
Small fix, add legacy class
