### 2019-06-14 7464a04
Changed generate_decay to nbody_decay shortcut.
