### 2019-06-17 b4b6fe5
Add deprecation warning
