### 2019-06-17 262632c
Add a small docstring in Particle
