+++ phasespace/phasespace.py
@@ -580,6 +580,7 @@ class Particle:
     Renamed to GenParticle.
 
     """
+
     def __init__(self):
         raise NameError("'Particle' has been renamed to 'GenParticle'. Please update your code accordingly."
                         "For more information, see: https://github.com/zfit/phasespace/issues/22")
