### 2019-06-26 #1490 @stuartcampbell
Hi Jack, (@poulson),  I don't really know anything about particle physics, my background is in condensed matter physics and neutron/x-ray scattering.  But if you still want me to, I am willing to have a look.
