### 2019-07-10 #1490 @vyasr
Apologies for not responding sooner. This package is unfortunately outside my realm of expertise, the particle simulations that I am familiar with are typically in soft condensed matter so I don't really know particle physics. I'm happy to provide code review, but if you'd like me to review I think it would be good to have a second reviewer who has some expertise in the field
