### 2019-07-10 #1490 @stuartcampbell
I'd be happy to do it , but I would assume that the answers it gives are correct.
