### 2019-07-11 #1490 @apuignav
Hey, just to point out that the tests check the physics answers against the accepted standard in particle physics, [TGenPhaseSpace](https://root.cern.ch/doc/master/classTGenPhaseSpace.html), and the [RapidSim](https://github.com/gcowan/RapidSim) software (extensively used in peer-reviewed publications), which makes use of TGenPhaseSpace for complex decay chains.
