### 2019-07-16 a3c25ff
Fix typography of paper
