### 2019-07-16 #1570 @apuignav
Fixed a couple of typography issues, now it's good to go!
