+++ phasespace/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = """Albert Puig Navarro"""
 __email__ = 'albert.puig@cern.ch'
-__version__ = '1.0.2'
+__version__ = '1.0.2dev0'
 __maintainer__ = "zfit"
 
 __credits__ = ["Jonas Eschle <jonas.eschle@cern.ch>"]
