### 2019-08-12 #1570 @mdoucet
I'm done with my review. The code does what it says it does. I could run the code, generate decays and check that the 4-momentum vectors and invariant mass of the results made sense. I only wish that functionality to deal with decay widths, along the lines of the helper code in `tests/helpers/decays.py`, would be part of the main code.
