### 2019-08-19 90cad56
Moved comparison unittest files to cloud storage
