### 2019-08-21 e0fff21
Removed log for CI
