### 2019-08-21 7740292
Fix CI test download in wrong directory
