### 2019-08-21 #1570 @mayou36
> * I don't think most of the elements added to setup_requires are actually necessary for users (e.g. sphinx, bumpversion, wheel, twine, etc). I would remove those that the user shouldn't actually need. Developers can install those as needed.

I agree, removed

> * A few tests currently fail because they're referring to nonexistent files. Based on your documentation, these tests (in `test_physics.py`) are the ones that are meant to test against external codes. If the outputs are being stored as artifacts directly on your testing servers or something similar, I would suggest using e.g. `pytest.mark.skip` if not on CI to avoid users seeing a failure.
> * Although they are internal methods, they're doing the bulk of the work it seems like, so I would like to see some documentation of the `_generate` and `_recursive_generate` functions for the sake of future developers.

Fixed the CI (there was a problem with uploading the large files to the CI)
