### 2019-08-22 #1570 @mayou36
> I'm done with my review.

Thanks a lot for that!

> I only wish that functionality to deal with decay widths, along the lines of the helper code in `tests/helpers/decays.py`, would be part of the main code.

This is intentionally not part of the package. phasespace aims to only provide the decay mechanics and leaves other things to other libraries, since there are enough around that are quite suitable for that. E.g. using `tensorflow_probability` with it's distributions and the `particle`-package for the particles characteristics works already well. And there is an example also provided in the docs that can be copy-pasted basically.

So from our point of view, the additional maintenance and design effort is simply not worth it.
