### 2019-08-27 #1570 @apuignav
@vyasr I think it makes sense, although the documentation is mainly included in the "parent" methods. We'll add docs to clarify better so it's more developer-friendly.
