### 2019-08-28 89e4580
Added missing docs.
