### 2019-08-29 #1570 @vyasr
Looks good! I'm all set with my review now.

One minor tip, the docstrings are pretty inconsistent with using `tensor`, `Tensor`, and `tf.Tensor`. If you consistently refer to `tensorflow.Tensor` using the fully qualified name (it might be possible to use aliases, I haven't tried) you can take advantage of [Intersphinx](http://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html) to get links to show up in your documentation.
