### 2019-08-30 9e5a41a
Remove deploy requirements from setup requirements
