### 2019-08-30 fcf1412
Separate tensor and tf.Tenso
