### 2019-09-03 #1570 @poulson
:wave: @openjournals/joss-eics, this submission looks ready to be accepted.
