### 2019-09-03 #1570 @labarba
Before continuing, @apuignav, please edit the metadata of the Zenodo archive so the author list and title matches the JOSS paper.
