### 2019-09-03 #1570 @apuignav
Should be done. I don't know why zenodo deletes all the info whenever a new release happens :-(
