+++ setup.py
@@ -27,7 +27,9 @@ if not requirements_dev.count("") == 1 or requirements_dev.index("") == 0:
                       "requirements have to be separated by one blank line.")
 requirements_dev_split = requirements_dev.index("")
 
-# setup_requirements = requirements_dev[:requirements_dev_split]
+setup_requirements = ["pip>9",
+                      "setuptools_scm",
+                      "setuptools_scm_git_archive"]
 test_requirements = requirements_dev[requirements_dev_split + 1:]  # +1: skip empty line
 
 setup(
@@ -57,6 +59,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/zfit/phasespace',
-    version='1.0.2',
+    use_scm_version=True,
     zip_safe=False,
 )
