+++ phasespace/__init__.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
 """Top-level package for TensorFlow PhaseSpace."""
+from pkg_resources import get_distribution
 
 __author__ = """Albert Puig Navarro"""
 __email__ = 'albert.puig@cern.ch'
-__version__ = '1.0.2dev0'
+__version__ = get_distribution(__name__).version
 __maintainer__ = "zfit"
 
 __credits__ = ["Jonas Eschle <jonas.eschle@cern.ch>"]
