+++ .travis.yml
@@ -1,15 +1,12 @@
-os: linux
 language: python
+os:
+    - linux
 python:
-  - 3.6
+    - 3.6
+    - 3.7
 
 cache: pip
 
-matrix:
-  include:
-    - python: 3.7
-      dist: xenial
-      sudo: true
 before_install:
   # get miniconda and install
   - wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O anaconda.sh;
@@ -31,6 +28,7 @@ install:
 script:
   - pytest --cov-branch --cov=phasespace tests
 after_success:
+  - pip install coveralls
   - coveralls
 
 deploy:
