+++ HISTORY.rst
@@ -1,0 +1,28 @@
+*********
+Changelog
+*********
+
+Develop
+=======
+
+Major Features and Improvements
+-------------------------------
+
+Behavioral changes
+------------------
+
+Bug fixes and small changes
+---------------------------
+
+
+Requirement changes
+-------------------
+
+
+Thanks
+------
+
+1.0.3 (13-10-2019)
+===================
+
+Release to conda-forge, thanks to Chris Burr
