### 2019-10-13 ed1966a
Adjust requirements below TF 2.0
