+++ requirements.txt
@@ -1,2 +1,2 @@
-tensorflow>=1.12.0
-tensorflow_probability
+tensorflow>=1.12.0,<2
+tensorflow_probability<0.8
