+++ phasespace/__init__.py
@@ -4,11 +4,11 @@
 from pkg_resources import get_distribution
 
 __author__ = """Albert Puig Navarro"""
-__email__ = 'albert.puig@cern.ch'
+__email__ = 'apuignav@gmail.com'
 __version__ = get_distribution(__name__).version
 __maintainer__ = "zfit"
 
-__credits__ = ["Jonas Eschle <jonas.eschle@cern.ch>"]
+__credits__ = ["Jonas Eschle <Jonas.Eschle@cern.ch>"]
 
 __all__ = ['nbody_decay', 'GenParticle']
 
