+++ HISTORY.rst
@@ -22,7 +22,12 @@ Requirement changes
 Thanks
 ------
 
+
 1.0.3 (13-10-2019)
-===================
+==========================
+
+
+Major Features and Improvements
+-------------------------------
 
 Release to conda-forge, thanks to Chris Burr
