+++ setup.cfg
@@ -1,17 +1,3 @@
-[bumpversion]
-current_version = 0.9.0
-commit = True
-tag = True
-tag_name = {new_version}
-
-[bumpversion:file:setup.py]
-search = version='{current_version}'
-replace = version='{new_version}'
-
-[bumpversion:file:phasespace/__init__.py]
-search = __version__ = '{current_version}'
-replace = __version__ = '{new_version}'
-
 [bdist_wheel]
 universal = 1
 
