+++ HISTORY.rst
@@ -23,7 +23,7 @@ Thanks
 ------
 
 
-1.0.3 (13-10-2019)
+1.0.4 (13-10-2019)
 ==========================
 
 
