### 2019-10-16 df60582
Add conda installation instruction
