### 2019-10-29 880681b
minor changes in paper
