### 2019-10-29 #1570 @danielskatz
Can you merge some minor changes in the paper: https://github.com/zfit/phasespace/pull/32 ?
