### 2019-10-29 #1570 @danielskatz
👋 @arfon - note that we have a problem with the title in the xml, due to the greek character, I assume. Your thoughts?
