### 2019-10-29 #1570 @arfon
Hrm, looks like Whedon is stripping this out. Feel free to accept here and I'll update the Crossref metadata manually after you accept.
