### 2019-10-29 #1570 @danielskatz
@arfon - please do your crossref-latex magic :)  And then please close this when everything seems to be right
