### 2019-10-29 #1570 @arfon
> @arfon - please do your crossref-latex magic :) And then please close this issue when everything seems to be right

OK, fixed up the metadata and opened https://github.com/openjournals/whedon/issues/59 to track this.
